+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Updated
 1. Reworked command line arguments to specify ACL table using --table:ACL option.
+2. Reworked sqlite3 implementation to replace the UPSERT with an ON CONFLICT REPLACE
+   clause in the CREATE TABLE.
 
 
 ## [0.8.5](https://github.com/uhppoted/uhppoted-app-db/releases/tag/v0.8.5) - 2023-06-13